RA-1244 - :construction: Splitting out VacancyMapper into separate mappers for Apprenticeship, Traineeship
